# api/openapi.yaml
openapi: 3.0.3
info:
  title: ContactBookAPI
  version: 1.0.0
  description: ContactBook API

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorObject:
      type: object
      properties:
        step:
          type: string
        error:
          type: string
      required:
        - step
        - error

    AuthObject:
      type: object
      properties:
        token:
          type: string
        expires:
          type: string
          format: date-time
      required:
        - token
        - expires

    SelectorObject:
      type: object
      properties:
        offset:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
      required:
        - limit

    PreloadObject:
      type: object
      properties:
        enabled:
          type: boolean
        primaryOnly:
          type: boolean
      required:
        - enabled

    PhoneObject:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        isPrimary:
          type: boolean
      required:
        - id
        - phone

    ContactObject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        birthday:
          type: string
          format: date
        note:
          type: string
        phones:
          type: array
          items:
            $ref: '#/components/schemas/PhoneObject'
      required:
        - id
        - name
        - birthday

    AddPhoneObject:
      type: object
      properties:
        phone:
          type: string
        isPrimary:
          type: boolean
      required:
        - phone

    AuthRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
          format: password
      required:
        - login
        - password

    AuthResponse:
      type: object
      properties:
        auth:
          $ref: '#/components/schemas/AuthObject'
      required:
        - auth

    RegisterRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
          format: password
      required:
        - login
        - password

    RegisterResponse:
      type: object
      properties:
        auth:
          $ref: '#/components/schemas/AuthObject'
      required:
        - auth

    GetContactsRequest:
      type: object
      properties:
        selector:
          $ref: '#/components/schemas/SelectorObject'
      required:
        - selector

    GetContactsResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/ContactObject'
      required:
        - contacts

    AddContactRequest:
      type: object
      properties:
        name:
          type: string
        birthday:
          type: string
          format: date
        note:
          type: string
        initialPhones:
          type: array
          items:
            $ref: '#/components/schemas/AddPhoneObject'
      required:
        - name
        - birthday
        - initialPhones

    AddContactResponse:
      type: object
      properties:
        contact:
          $ref: '#/components/schemas/ContactObject'
      required:
        - contact

    GetContactRequest:
      type: object
      properties:
        preload:
          $ref: '#/components/schemas/PreloadObject'
        withNote:
          type: boolean

    GetContactResponse:
      type: object
      properties:
        contact:
          $ref: '#/components/schemas/ContactObject'
      required:
        - contact

    UpdateContactRequest:
      type: object
      properties:
        name:
          type: string
        birthday:
          type: string
          format: date
        note:
          type: string
        
    UpdateContactResponse:
      type: object
      properties:
        contact:
          $ref: '#/components/schemas/ContactObject'
      required:
        - contact

    DeleteContactResponse:
      type: object

paths:
  /api/v1/auth:
    post:
      summary: User authentication
      operationId: authUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '401':
          description: Invalid login or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
  /api/v1/auth/register:
    post:
      summary: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
  /api/v1/contacts:
    post:
      summary: Retrieve a list of contacts
      operationId: getContacts
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetContactsRequest'
      responses:
        '200':
          description: Contacts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetContactsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'

  /api/v1/contact:
    post:
      summary: Add a new contact
      operationId: addContact
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddContactRequest'
      responses:
        '200':
          description: Contact added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddContactResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'

  /api/v1/contact/{contact_id}:
    post:
      summary: Retrieve a contact
      operationId: getContact
      security:
        - BearerAuth: []
      parameters:
        - name: contact_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetContactRequest'
      responses:
        '200':
          description: Contact retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetContactResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'

    patch:
      summary: Update a contact
      operationId: updateContact
      security:
        - BearerAuth: []
      parameters:
        - name: contact_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactRequest'
      responses:
        '200':
          description: Contact updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateContactResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'

    delete:
      summary: Delete a contact
      operationId: deleteContact
      security:
        - BearerAuth: []
      parameters:
        - name: contact_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Contact deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteContactResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'